// src/app/modules/clients/model/HomeOwernshipModel.ts
import { MlcModel } from './common/mlc.model';

export class HomeOwernship implements MlcModel {
  name: string;
  propertyorhomeaddress: string;
  ownership: string;
  nameofselleroragent: string;
  mortgageorloanaccountinstitutionname: string;
  branchnameorlocation: string;
  accountorloanname: string;
  namesorsignatoriesonloan: string;
  loanterm: string;
  loantype: string;
  customernumber: string;
  loanexpiryorrenewaldate: string;
  notes: string;
}